Fix IDEUI-293 Enhance notification behaviour
Some bugs were introduced by commit 25d32d on this issue. The text was
no more updatable and the in progress icon was always overriden by the
success one. Fix all thi things.

Moreover, each setter notify the observers, triggering a sequence of
refresh. The update method has been added to update a bunch of
parameters in a row and notify observers after.