Fix IDEUI-316 Trigger zombie runners
When launching an application, the Run button becomes disabled only
when the app is considered running, so you can click multiple times
before that and ‘launch’ multiple runners. They don't really work but
are like staying in background in the events view.

The new flag isAnyAppLaunched in RunController is now setted just after
launch and used to enable/disable RunAction, preventing multi-run
trigerring.